bobcard2.0/requirements.txt
Django==2.1.3
wheel==0.32.3
gunicorn==19.9.0
psycopg2==2.7.6.1
django-qr-code
django-background-tasks